Rename package to lowercase
