+++ example.py
@@ -1,13 +1,13 @@
 #!/usr/bin/python
 import time
 
-from MS5837 import ms5837
+from ms5837 import ms5837
 
 sensor = ms5837.MS5837_30BA() # Default I2C bus is 1 (Raspberry Pi 3)
 #sensor = ms5837.MS5837_30BA(0) # Specify I2C bus
 #sensor = ms5837.MS5837_02BA()
 #sensor = ms5837.MS5837_02BA(0)
-#sensor = ms5837.MS5837(model=ms5837.MS5837_MODEL_30BA, bus=0) # Specify model and bus
+#sensor = ms5837.ms5837(model=ms5837.MS5837_MODEL_30BA, bus=0) # Specify model and bus
 
 # We must initialize the sensor before reading it
 if not sensor.init():
